+++ fuzzing/fuzz_levenshtein_editops.cpp
@@ -8,10 +8,10 @@
 #include <string>
 
 void validate_editops(const std::basic_string<uint8_t>& s1, const std::basic_string<uint8_t>& s2,
-                      int64_t score, int64_t score_hint = std::numeric_limits<int64_t>::max())
+                      size_t score, size_t score_hint = std::numeric_limits<size_t>::max())
 {
     rapidfuzz::Editops ops = rapidfuzz::levenshtein_editops(s1, s2, score_hint);
-    if (static_cast<int64_t>(ops.size()) == score && s2 != rapidfuzz::editops_apply<uint8_t>(ops, s1, s2))
+    if (ops.size() == score && s2 != rapidfuzz::editops_apply<uint8_t>(ops, s1, s2))
         throw std::logic_error("levenshtein_editops failed");
 }
 
@@ -23,7 +23,7 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t* data, size_t size)
     /* hirschbergs algorithm is only used for very long sequences which are apparently not generated a lot by
      * the fuzzer */
     for (int i = 0; i < 10; i++) {
-        int64_t score = rapidfuzz_reference::levenshtein_distance(s1, s2);
+        size_t score = rapidfuzz_reference::levenshtein_distance(s1, s2);
         validate_editops(s1, s2, score);
         validate_editops(s1, s2, score, 64);
         validate_editops(s1, s2, score, score != 0 ? score - 1 : 0);
